+++ WinDesktopApp/FontShader.h
@@ -11,7 +11,7 @@ class CFontShader : private CTextureShader
 public:
 	virtual bool Initialize(ID3D11Device* pDevice, HWND hwnd) override;
 	virtual void Shutdown() override;
-	bool Render(ID3D11DeviceContext* pDeviceContext, int nIndexCount
+	bool Render(ID3D11DeviceContext* pDeviceContext, size_t nIndexCount
 		, XMMATRIX matrWorld, XMMATRIX matrView, XMMATRIX matrProjection
 		, ID3D11ShaderResourceView* pTextureSRV, XMFLOAT4 color);
 
@@ -19,7 +19,7 @@ protected:
 	bool InitializeShader(ID3D11Device* pDevice, HWND hwnd, const WCHAR* szFileNameVS, const WCHAR* szFileNamePS);
 	bool SetShaderParameters(ID3D11DeviceContext* pDeviceContext, XMMATRIX matrWorld
 		, XMMATRIX matrView, XMMATRIX matrProjection, ID3D11ShaderResourceView* pTexture, XMFLOAT4 color);
-	virtual void RenderShader(ID3D11DeviceContext* pDeviceContext, int nIndexCount) override;
+	virtual void RenderShader(ID3D11DeviceContext* pDeviceContext, size_t nIndexCount) override;
 	virtual void ShutdownShader() override;
 
 protected:
