+++ WinDesktopApp/LightShader.h
@@ -14,7 +14,7 @@ class CLightShader : public CTextureShader
 
 public:
 	virtual bool Initialize(ID3D11Device* pDevice, HWND hwnd);
-	bool Render(ID3D11DeviceContext* pDeviceContext, int nIndexCount, XMMATRIX matrWorld
+	bool Render(ID3D11DeviceContext* pDeviceContext, size_t nIndexCount, XMMATRIX matrWorld
 		, XMMATRIX matrView, XMMATRIX matrProjection, ID3D11ShaderResourceView* pTexture, XMFLOAT3 lightDirection, XMFLOAT4 diffuseColor);
 
 protected:
