+++ WinDesktopApp/TextureShader.h
@@ -16,13 +16,13 @@ public:
 
 	virtual bool Initialize(ID3D11Device* pDevice, HWND hwnd);
 	virtual void Shutdown();
-	bool Render(ID3D11DeviceContext* pDeviceContext, int nIndexCount, XMMATRIX matrWorld
+	bool Render(ID3D11DeviceContext* pDeviceContext, size_t nIndexCount, XMMATRIX matrWorld
 		, XMMATRIX matrView, XMMATRIX matrProjection, ID3D11ShaderResourceView* pTexture);
 
 protected:
 	virtual bool InitializeShader(ID3D11Device* pDevice, HWND hwnd, const WCHAR* szPathVS, const WCHAR* szPathPS);
 	virtual void ShutdownShader() override;
-	virtual void RenderShader(ID3D11DeviceContext* pDiviceContext, int nIndexCount);
+	virtual void RenderShader(ID3D11DeviceContext* pDiviceContext, size_t nIndexCount);
 
 	virtual bool CreateSamplerState(ID3D11Device* pDevice);
